feat: add polygon parts schema and types for intersection api (#193)
